fix: 코드리뷰 반영
타입변환 변경. 필드값 세팅 변경

+++ src/main/java/org/swyp/weddy/domain/auth/service/AuthService.java
@@ -111,7 +111,7 @@ public class AuthService {
     public UserResponse getUserInfo() {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         Map<String, Object> principal = (Map<String, Object>) authentication.getPrincipal();
-        Member member = memberMapper.selectByMemberId(Long.valueOf((Integer) principal.get("id")));
+        Member member = memberMapper.selectByMemberId((Long) principal.get("id"));
         return UserResponse.from(member);
     }
 }
